refactor(folder structure): updated the folders to include each data structure in its own folder
