+++ src/LinkedList/LinkedList.java
@@ -1,4 +1,4 @@
-package packages.LinkedList;
+package LinkedList;
 
 import java.util.HashSet;
 import java.util.NoSuchElementException;
